Move snippet converter to pxtc
